% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_arrow}
\alias{add_arrow}
\title{Add arrow}
\usage{
add_arrow(
  file,
  x0,
  y0,
  x1 = x0,
  y1 = y0,
  arrow_col = "red",
  arrow_width = 2L,
  head_angle = 30,
  head_length = 0.2
)
}
\arguments{
\item{file}{The name of a file (without extension) to add to}

\item{x0}{starting x coordinate}

\item{y0}{starting y coordinate}

\item{x1}{end x coordinate}

\item{arrow_col}{colour of the line}

\item{arrow_width}{width of the line}

\item{head_angle}{angle of the arrow head}

\item{head_length}{length of the arrow head}

\item{x2}{end y coordinate}
}
\value{
NULL, saves image
}
\description{
Add arrow
}
