% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_circle}
\alias{add_circle}
\title{Add circle}
\usage{
add_circle(file, x0, y0, radius, col = "red", width = 2L, fill_col = NA)
}
\arguments{
\item{file}{The name of a file (without extension) to add to}

\item{x0}{center x coordinate}

\item{y0}{center y coordinate}

\item{radius}{size of the circle}

\item{col}{colour of the border}

\item{width}{width of the border}

\item{fill_col}{fill colour}
}
\value{
NULL, saves image
}
\description{
Add circle
}
