% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_line}
\alias{add_line}
\title{Add line}
\usage{
add_line(file, x0, y0, x1 = x0, y1 = y0, line_col = "red", line_width = 2L)
}
\arguments{
\item{file}{The name of a file (without extension) to add to}

\item{x0}{starting x coordinate}

\item{y0}{starting y coordinate}

\item{x1}{end x coordinate}

\item{line_col}{colour of the line}

\item{line_width}{width of the line}

\item{x2}{end y coordinate}
}
\value{
NULL, saves image
}
\description{
Add line
}
