% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_rect}
\alias{add_rect}
\title{Add rectangle}
\usage{
add_rect(
  file,
  x0,
  y0,
  x1,
  y1,
  border_col = "red",
  border_width = 2,
  fill_col = NA
)
}
\arguments{
\item{file}{The name of a file (without extension) to add to}

\item{x0}{left side of the rectangle}

\item{y0}{bottom of the rectangle}

\item{x1}{right side of the rectangle}

\item{y1}{top of the rectangle}

\item{border_col}{colour of the border}

\item{border_width}{width of the border}

\item{fill_col}{colour to will the rectangle}
}
\value{
NULL, saves image
}
\description{
Add rectangle
}
