% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_text}
\alias{add_text}
\title{Add text}
\usage{
add_text(
  file,
  x0,
  y0,
  text,
  font_col = "red",
  font_size = 1,
  font_style = 1L,
  justify_x = 0.5,
  justify_y = NA
)
}
\arguments{
\item{file}{The name of a file (without extension) to add to}

\item{x0}{x position}

\item{y0}{y position}

\item{text}{the text to add}

\item{font_col}{the colour of the text}

\item{font_size}{the size of the text}

\item{font_style}{the style of the text (see ?text)}

\item{justify_x}{the x justification of the text (see ?text)}

\item{justify_y}{the y justification of the text (see ?text)}
}
\value{
NULL, saves image
}
\description{
Add text
}
