% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{bitmap2png}
\alias{bitmap2png}
\title{Converts a bitmap image to PNG using}
\usage{
bitmap2png(path)
}
\arguments{
\item{path}{Path to the bitmap image.}
}
\value{
The result of the system call.
}
\description{
Converts a bitmap image to PNG using
}
\examples{
\dontrun{
bitmap2png("path/to/image.bmp")
}

}
