% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{Capture Screenshot}
\usage{
capture(file, grid = getOption("screenshot.dev.grid", FALSE))
}
\arguments{
\item{file}{The name of the file}

\item{grid}{optional, if TRUE then a grid is overlaid to help with
positioning}
}
\value{
NULL, generates a file called 'file.png'
}
\description{
Capture Screenshot
}
