% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard2bitmap}
\alias{clipboard2bitmap}
\title{Save clipboard image to temporary BMP file}
\usage{
clipboard2bitmap()
}
\value{
Path to the temporary BMP file.
}
\description{
This function works only on windows.
This function saves the image currently in the clipboard
to a temporary BMP file.
}
\examples{
\dontrun{
clipboard2bitmap()
}
}
