% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clipboard_sample}
\alias{clipboard_sample}
\title{Sample data of clipboard bitmap array}
\format{
raw data of bitmap array with 246292 length
}
\usage{
clipboard_sample
}
\description{
Sample data of clipboard bitmap array
}
\examples{
data(clipboard_sample)

}
\keyword{datasets}
