% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_image.R
\name{compare_table}
\alias{compare_table}
\title{Compare values within tow arrays or matrices.
Helper function for \code{locate_ndl_in_hay()}.}
\usage{
compare_table(ndl_mt, hay_mt)
}
\arguments{
\item{ndl_mt, hay_mt}{A matrix.}
}
\value{
A tibble.
}
\description{
Compare values within tow arrays or matrices.
Helper function for \code{locate_ndl_in_hay()}.
}
\examples{
val <- seq(from = 0, to = 1, by = 0.1)
mt_1 <- matrix(sample(val,  20, replace = TRUE))
mt_2 <- matrix(sample(val, 100, replace = TRUE))
compare_table(mt_1, mt_2)

}
