% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_image.R
\name{count_val_freq}
\alias{count_val_freq}
\title{Helper function for \code{compare_table()}.}
\usage{
count_val_freq(mt, colname)
}
\arguments{
\item{mt}{A numeric matrix or array.}

\item{colname}{A string of name for count.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A dataframe.
}\if{html}{\out{</div>}}
}
\description{
Helper function for \code{compare_table()}.
}
\examples{
mt <- sample(1:10, 30, replace = TRUE)
count_val_freq(mt, "freq")

}
