% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl}
\alias{crawl}
\title{Crawl a directory for screenshots}
\usage{
crawl(dir)
}
\arguments{
\item{dir}{the directory to crawl}
}
\value{
NULL, saves images
}
\description{
Crawl a directory for screenshots
}
