% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{create_header}
\alias{create_header}
\title{Create a BMP header}
\usage{
create_header(clipboard)
}
\arguments{
\item{clipboard}{A raw vector of the clipboard contents.}
}
\value{
A raw vector of the BMP header.
}
\description{
Create a BMP header
}
\examples{
data(clipboard_sample)
create_header(clipboard_sample)

}
