% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Crop image}
\usage{
crop(file, x0 = 0, x1 = img_info$width[1], y0 = 0, y1 = img_info$height[1])
}
\arguments{
\item{file}{The name of a file (without extension) to add to}

\item{x0}{left cropping point}

\item{x1}{right cropping point}

\item{y0}{top cropping point}

\item{x2}{bottom cropping point}
}
\value{
NULL, saves image
}
\description{
Crop image
}
