% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_size}
\alias{display_size}
\title{Get the size of the display.}
\usage{
display_size()
}
\value{
A list with two elements, \code{width} and \code{height}, which are the width
and height of the display.
}
\description{
This function works only on windows.
}
\examples{
\dontrun{
display_size()
}

}
