% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{get_clipboard_image}
\alias{get_clipboard_image}
\title{Retrieves the image from the clipboard}
\usage{
get_clipboard_image()
}
\value{
A raw vector containing the image data.
}
\description{
This function works only on windows.
}
\examples{
\dontrun{
get_clipboard_image()
}

data(clipboard_sample)
head(clipboard_sample, 100)
header <- create_header(clipboard_sample)
image_data <- c(header, clipboard_sample)
path <- fs::path_temp(ext = "bmp")
save_bmp(image_data, path)
 # shell.exec(path)
fs::file_delete(path)

}
