% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locat_image.R
\name{hay2needle}
\alias{hay2needle}
\title{Cut off a part of image from a whole image.}
\usage{
hay2needle(haystack_image, pos_x, pos_y, w = 50, h = 20)
}
\arguments{
\item{haystack_image}{An image of cimg.}

\item{pos_x, pos_y}{A numeric to indicate the top left corner of cutting image.
When NULL, position will be randomly sampled.}

\item{w, h}{A numeric for width or height of the cutting image.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          An image of cimg object.
}\if{html}{\out{</div>}}
}
\description{
Cut off a part of image from a whole image.
}
\examples{
library(imager)
haystack_image <- imager::load.example("parrots")
needle_image <- hay2needle(haystack_image, 200, 250, 100, 50)
layout(c(1:2))
plot(haystack_image)
plot(needle_image)

}
