% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_image.R
\name{image2gray_matrix}
\alias{image2gray_matrix}
\title{Convert cimg class into grayscale xy matrix.
Helper function for \code{locate_image()}.
Use grayscale to Speed up and to simplify code.}
\usage{
image2gray_matrix(img)
}
\arguments{
\item{img}{A cimg object.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ An xy dimensional matrix.
}\if{html}{\out{</div>}}
}
\description{
Convert cimg class into grayscale xy matrix.
Helper function for \code{locate_image()}.
Use grayscale to Speed up and to simplify code.
}
