% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_image.R
\name{index2xy}
\alias{index2xy}
\title{Convert array index into xy location in matrix.
Helper function for \code{locate_ndl_in_hay()}.}
\usage{
index2xy(index, nrow)
}
\arguments{
\item{index, nrow}{A numeric.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       A numeric pair of xy location.
}\if{html}{\out{</div>}}
}
\description{
Convert array index into xy location in matrix.
Helper function for \code{locate_ndl_in_hay()}.
}
\examples{
nrow <- 4
matrix(1:12, nrow = nrow)
purrr::map(1:12, index2xy, nrow = nrow)

}
