% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{install_screenshot}
\alias{install_screenshot}
\title{Install command line screenshot for Windows.}
\usage{
install_screenshot(dir)
}
\arguments{
\item{dir}{A string of direcotory to be installed.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A string of installed dir.
}\if{html}{\out{</div>}}
}
\description{
Codes are from URL shown below.
https://superuser.com/questions/75614/take-a-screen-shot-from-command-line-in-windows#answer-1751844
}
\examples{
\donttest{
library(fs)
dir <- fs::path_package("screenshot")
# dir <- "SET_YOUR DIRECTORY"
install_screenshot(dir)
}

}
