% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{reset_transparent}
\alias{reset_transparent}
\title{Reset transparent background of an image}
\usage{
reset_transparent(path)
}
\arguments{
\item{path}{A character string specifying the file path to the image to be processed.}
}
\value{
The file path to the processed image.
}
\description{
Reset transparent background of an image
}
\examples{
\dontrun{
path <- "directory/image.png"
reset_transparent(path)
}

}
