% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{save_clipboard_image}
\alias{save_clipboard_image}
\title{Saves an image from the clipboard to a file}
\usage{
save_clipboard_image(path = "", reset_transparent = TRUE)
}
\arguments{
\item{path}{Optional path to save the image to.
If not specified, a temporary file will be created.}

\item{reset_transparent}{A logical. If true, the image of
transparent color will be removed.}
}
\value{
The path to the saved image file.
}
\description{
This function works only on windows.
}
\examples{
\dontrun{
# Save the image from the clipboard to a file
save_clipboard_image("clipboard_image.png")
}

}
