% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Take a screenshot.}
\usage{
screenshot(file = "", bin_dir = "", quote = FALSE)
}
\arguments{
\item{file}{A string for file name of screenshot.}

\item{bin_dir}{A string for directory name of screenshot.exe on Win.}

\item{quote}{Deprecated and ignored. The command is always quoted.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A file name of screenshot. When "", screenshot will be saved in a tempral directory.
}\if{html}{\out{</div>}}
}
\description{
Need to install screenshot.exe on Win by install_screenshot().
}
\details{
The command is always quoted, so that a path with a space
(such as "C:/Program Files/R/R-4.5.1/library/screenshot") works.
}
\examples{
if(interactive()){

sc <- screenshot()
if(sc != ""){
  sc_image <- imager::load.image(sc)
  plot(sc_image)
}

}

}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   install_screenshot()
}\if{html}{\out{</div>}}
}
