% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot_exists}
\alias{screenshot_exists}
\title{Find screenshot exec file.}
\usage{
screenshot_exists(bin_dir = "")
}
\arguments{
\item{bin_dir}{A string for directory name screenshot.exe exec file.
No need on Mac and Linux.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A logical.
}\if{html}{\out{</div>}}
}
\description{
Find screenshot exec file.
}
\examples{
screenshot_exists()

}
