% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_image.R
\name{crop_image}
\alias{crop_image}
\alias{hay2needle}
\title{Cut off a part of image from a whole image.}
\usage{
crop_image(image, pos_x, pos_y, w = 50, h = 20)

hay2needle(image, pos_x, pos_y, w, h)
}
\arguments{
\item{image}{An image of cimg.}

\item{pos_x, pos_y}{A numeric to indicate the top left corner of cutting image.
When NULL, position will be randomly sampled.}

\item{w, h}{A numeric for width or height of the cutting image.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          An image of cimg object.
}\if{html}{\out{</div>}}
}
\description{
Cut off a part of image from a whole image.
}
\examples{
image <- imager::load.example("parrots")
croped_image <- crop_image(image, 200, 250, 100, 50)
layout(c(1:2))
plot(image)
plot(croped_image)

}
