% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_corner}
\alias{display_corner}
\title{Get display corner of screen}
\usage{
display_corner(corner = "bottom_left", width = 600, height = 600)
}
\arguments{
\item{corner}{A string to specify a corner of the display.
"top_left", "top_right", "bottom_left", or "bottom_right".}

\item{width, height}{A integer to specify width or height of the corner.}
}
\value{
A numeric vector of length 4 representing the coordinates of the specified corner.
}
\description{
This function returns the coordinates of the specified corner of the display.
This function works only on windows.
}
\examples{
\dontrun{
display_corner("top_left", 800, 800)
}

}
