% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_image.R
\name{is_all_same}
\alias{is_all_same}
\title{Helper function for \code{locate_ndl_in_hay()}.}
\usage{
is_all_same(ndl_mt, hay_mt, base_xy)
}
\arguments{
\item{ndl_mt, hay_mt}{A matrix}

\item{base_xy}{A numeric pair of xy location.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A logical.
}\if{html}{\out{</div>}}
}
\description{
Helper function for \code{locate_ndl_in_hay()}.
}
