% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{save_clipboard_image}
\alias{save_clipboard_image}
\title{Saves an image from the clipboard to a file}
\usage{
save_clipboard_image(bin_dir, path = "")
}
\arguments{
\item{bin_dir}{Path to the directory containing the
\code{convert} binary (ImageMagick).}

\item{path}{Optional path to save the image to.
If not specified, a temporary file will be created.}
}
\value{
The path to the saved image file.
}
\description{
This function works only on windows.
}
\examples{
\dontrun{
# Save the image from the clipboard to a file
bin_dir <- "DIR/OF/CONVERT/BINARY"
save_clipboard_image(bin_dir, "clipboard_image.png")
}

}
