% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Take a screenshot.}
\usage{
screenshot(file = "")
}
\arguments{
\item{file}{A string for file name of screenshot.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A file name of screenshot. When "", screenshot will be saved in a tempral directory.
}\if{html}{\out{</div>}}
}
\description{
Need to install screenCapture on Win by install_screenshot().
}
\examples{
\donttest{
library(imager)
sc <- screenshot()
imager::load.image(sc)
}

}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{install_screenshot()
}\if{html}{\out{</div>}}
}
